text-generation/api/gemini/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
docstring_parser==0.16
exceptiongroup==1.2.2
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.8.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
numpy==2.2.4
packaging==24.2
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.1
pydantic_core==2.33.0
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
vertexai==1.71.1
websockets==15.0.1
